Reorganize manager modules into dedicated folder
