+++ src/game.js
@@ -4,19 +4,19 @@ import { SETTINGS } from '../config/gameSettings.js';
 import { GameLoop } from './gameLoop.js';
 import { InputHandler } from './inputHandler.js';
 import { CharacterFactory, ItemFactory } from './factory.js';
-import { EventManager } from './eventManager.js';
-import { CombatLogManager, SystemLogManager } from './logManager.js';
+import { EventManager } from './managers/eventManager.js';
+import { CombatLogManager, SystemLogManager } from './managers/logManager.js';
 import { CombatCalculator } from './combat.js';
 import { MapManager } from './map.js';
 import { MonsterManager, MercenaryManager, ItemManager, EquipmentManager, UIManager, VFXManager, SkillManager, SoundManager } from './managers/index.js';
 import { AssetLoader } from './assetLoader.js';
-import { MetaAIManager, STRATEGY } from './ai-managers.js';
-import { SaveLoadManager } from './saveLoadManager.js';
-import { LayerManager } from './layerManager.js';
-import { PathfindingManager } from './pathfindingManager.js';
-import { FogManager } from './fogManager.js';
-import { NarrativeManager } from './narrativeManager.js';
-import { TurnManager } from './turnManager.js';
+import { MetaAIManager, STRATEGY } from './managers/ai-managers.js';
+import { SaveLoadManager } from './managers/saveLoadManager.js';
+import { LayerManager } from './managers/layerManager.js';
+import { PathfindingManager } from './managers/pathfindingManager.js';
+import { FogManager } from './managers/fogManager.js';
+import { NarrativeManager } from './managers/narrativeManager.js';
+import { TurnManager } from './managers/turnManager.js';
 import { SKILLS } from './data/skills.js';
 import { EffectManager } from './managers/effectManager.js';
