+++ src/managers/ai-managers.js
