+++ src/managers/eventManager.js
