+++ src/managers/fogManager.js
