+++ src/managers/layerManager.js
