+++ src/managers/logManager.js
