+++ src/managers/narrativeManager.js
