+++ src/managers/pathfindingManager.js
