+++ tests/effectManager.test.js
@@ -1,5 +1,5 @@
 import { EffectManager } from '../src/managers/effectManager.js';
-import { EventManager } from '../src/eventManager.js';
+import { EventManager } from '../src/managers/eventManager.js';
 import { test, assert } from './helpers.js';
 
 console.log("--- Running EffectManager Tests ---");
