+++ tests/eventManager.integration.test.js
@@ -1,4 +1,4 @@
-import { EventManager } from '../src/eventManager.js';
+import { EventManager } from '../src/managers/eventManager.js';
 import { test, assert } from './helpers.js';
 
 console.log("--- Running EventManager Integration Tests ---");
